app/server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==14.1
websocket-client==1.7.0
sqlalchemy==2.0.36
pandas==2.1.4
numpy==1.26.4
scipy==1.13.1
statsmodels==0.14.2
scikit-learn==1.5.2
python-multipart==0.0.12
aiofiles==24.1.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
pykalman==0.9.5


